backend/app/api/endpoints/system.py
from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.orm import Session
import os
from loguru import logger

from ...core.database import SystemSetting, AgentConfig, SystemLog, get_db
from ...schemas.system import UpdateSettingRequest, TestConnectionRequest
import httpx

router = APIRouter()

@router.get("/settings", tags=["settings"])
async def get_settings(db: Session = Depends(get_db)):
    try:
        defaults = [
            {"key": "openai_model", "value": "gpt-4.1", "group": "llm"},
            {"key": "google_model", "value": "gemini-1.5-flash", "group": "llm"},
            {"key": "llm_provider", "value": os.getenv("LLM_PROVIDER", "gemini"), "group": "llm"},
            {"key": "openai_api_key", "value": os.getenv("OPENAI_API_KEY", ""), "group": "llm"},
            {"key": "google_api_key", "value": os.getenv("GEMINI_API_KEY", ""), "group": "llm"},
            {"key": "temperature", "value": "0.7", "group": "llm"},
            {"key": "max_tokens", "value": "4096", "group": "llm"},
            {"key": "realtime_streaming", "value": "true", "group": "llm"},
            {"key": "guardrails_input_scan", "value": "true", "group": "security"},
            {"key": "guardrails_output_validation", "value": "true", "group": "security"},
            {"key": "guardrails_pii_masking", "value": "true", "group": "security"},
            {"key": "guardrails_competitor_filter", "value": "true", "group": "security"}
        ]
        
        for d in defaults:
            existing = db.query(SystemSetting).filter(SystemSetting.key == d["key"]).first()
            if not existing:
                db.add(SystemSetting(**d))
        
        db.query(SystemSetting).filter(SystemSetting.key == "env_mode").delete()
        
        db.commit()
        
        settings = db.query(SystemSetting).all()
        return [{"key": s.key, "value": s.value, "group": s.group} for s in settings]
    except Exception as e:
        logger.error(f"Settings Error: {e}")
        db.rollback()
        raise HTTPException(status_code=500, detail="Failed to fetch settings")

@router.patch("/settings/{key}", tags=["settings"])
async def update_setting(key: str, request: UpdateSettingRequest, db: Session = Depends(get_db)):
    setting = db.query(SystemSetting).filter(SystemSetting.key == key).first()
    if not setting:
        raise HTTPException(status_code=404, detail="Setting not found")
    setting.value = request.value
    db.commit()
    return {"status": "success"}

@router.post("/settings/test-connection", tags=["settings"])
async def test_connection(request: TestConnectionRequest, db: Session = Depends(get_db)):
    try:
        if request.provider == "openai":
            async with httpx.AsyncClient() as client:
                res = await client.post(
                    "https://api.openai.com/v1/chat/completions",
                    headers={"Authorization": f"Bearer {request.api_key}"},
                    json={
                        "model": "gpt-3.5-turbo",
                        "messages": [{"role": "user", "content": "hi"}],
                        "max_tokens": 1
                    },
                    timeout=10.0
                )
                if res.status_code == 200:
                    return {"status": "success", "message": "Connection established with OpenAI."}
                else:
                    return {"status": "error", "message": f"OpenAI Error: {res.json().get('error', {}).get('message', 'Unknown error')}"}
        
        elif request.provider == "google":
            import google.generativeai as genai
            genai.configure(api_key=request.api_key)
            
            model_setting = db.query(SystemSetting).filter(SystemSetting.key == "google_model").first()
            model_name = model_setting.value if model_setting else "gemini-1.5-flash"
            
            if model_name.startswith("models/"):
                model_name = model_name.replace("models/", "", 1)
            
            model = genai.GenerativeModel(model_name)
            response = model.generate_content("hi")
            if response.text:
                return {"status": "success", "message": "Connection established with Google Gemini."}

        elif request.provider == "servicenow":
            instance_url = request.url or os.getenv("SERVICENOW_INSTANCE", "https://dev269222.service-now.com")
            username = request.username or os.getenv("SERVICENOW_USER", "admin")
            password = request.password or os.getenv("SERVICENOW_PASSWORD")
            
            if not password:
                return {"status": "error", "message": "ServiceNow password is required."}
            
            import requests
            from requests.auth import HTTPBasicAuth
            
            url = f"{instance_url.rstrip('/')}/api/now/table/incident"
            response = requests.get(
                url,
                auth=HTTPBasicAuth(username, password),
                headers={"Accept": "application/json"},
                params={"sysparm_limit": 1},
                timeout=10.0
            )
            
            if response.status_code == 200:
                return {"status": "success", "message": "Connection established with ServiceNow."}
            else:
                return {"status": "error", "message": f"ServiceNow Error: Status {response.status_code}"}

        elif request.provider == "tavily":
            api_key = request.api_key or os.getenv("TAVILY_API_KEY")
            if not api_key:
                return {"status": "error", "message": "Tavily API key is required."}
            
            async with httpx.AsyncClient() as client:
                res = await client.post(
                    "https://api.tavily.com/search",
                    json={
                        "api_key": api_key,
                        "query": "test",
                        "search_depth": "basic"
                    },
                    timeout=10.0
                )
                if res.status_code == 200:
                    return {"status": "success", "message": "Connection established with Tavily Search."}
                else:
                    return {"status": "error", "message": f"Tavily Error: {res.text}"}
            
        return {"status": "error", "message": "Unsupported provider"}
    except Exception as e:
        logger.error(f"Connection Test Error: {e}")
        return {"status": "error", "message": str(e)}

@router.get("/stats", tags=["system"])
async def get_stats(db: Session = Depends(get_db)):
    try:
        agents_count = db.query(AgentConfig).count()
        logs_count = db.query(SystemLog).count()
        errors_count = db.query(SystemLog).filter(SystemLog.status == "Failure").count()
        
        return {
            "agents": f"{agents_count} Agents",
            "throughput": "1,280/s",
            "uptime": "99.98%",
            "critical_tasks": str(errors_count),
            "memory_usage": "84%",
            "avg_latency": "12ms"
        }
    except Exception as e:
        logger.error(f"Stats Error: {e}")
        raise HTTPException(status_code=500, detail="Failed to fetch stats")

@router.get("/health", tags=["system"])
async def health():
    return {"status": "ok"}

@router.get("/integrations", tags=["system"])
async def get_integrations():
    openai_status = "Connected" if os.getenv("OPENAI_API_KEY") else "Disconnected"
    sn_status = "Live_Sync" if os.getenv("SERVICENOW_USER") and os.getenv("SERVICENOW_PASSWORD") else "Disconnected"
    tavily_status = "Connected" if os.getenv("TAVILY_API_KEY") else "Disconnected"
    
    return [
        {"id": "servicenow", "name": "ServiceNow Bridge", "status": sn_status, "endpoint": os.getenv("SERVICENOW_INSTANCE", "https://capgemini.service-now.com"), "latency": "12ms"},
        {"id": "tavily", "name": "Web Search Engine", "status": tavily_status, "latency": "45ms"},
        {"id": "database", "name": "PostgreSQL Neural Ledger", "status": "Connected", "latency": "2ms"}
    ]
