backend/requirements.txt
fastapi
uvicorn
pydantic-ai
langgraph
python-dotenv
openai
httpx
pydantic
websockets
sqlalchemy
mcp
requests
loguru
pysnow
pipecat-ai[openai,deepgram,cartesia,silero,google]==0.0.98
langchain-core
typing-extensions
psycopg2-binary
qdrant-client
langchain-openai
langchain-qdrant
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
docling
sse-starlette
tavily-python
langchain-google-genai
google-generativeai
paho-mqtt
